Refactor react router stuff into separate components
