+++ app/javascript/packs/containers/BuildingContainer.jsx
@@ -12,10 +12,4 @@ class BuildingContainer extends React.Component {
   }
 }
 
-function mapStateToProps(state) {
-  return {}
-}
-
-export default connect(
-  mapStateToProps
-)(BuildingContainer)
+export default BuildingContainer
