+++ app/views/buildings/show.html.erb
@@ -4,11 +4,4 @@
       <strong>Name:</strong>
       <%= @building.name %>
     </p>
-
-    <p>
-      <strong>Email:</strong>
-      <%= @building.contact_email %>
-    </p>
-
-
 </p>
