+++ app/views/portfolios/show.html.erb
@@ -6,9 +6,4 @@
   <strong>Asset manager: </strong><%= @portfolio.asset_manager.first_name %>
 </p>
 
-<!--<h3>Buildings</h3>-->
-<!--<% @portfolio.buildings.each do |b| %>-->
-  <!--<p><%= b.name %> | <%= link_to('Details', building_path(b)) %></p>-->
-<!--<% end %>-->
-
-<%= javascript_pack_tag 'containers/AssetManagerContainer.js' %>
+<%= javascript_pack_tag 'containers/RootContainer.js' %>
